settin up routing
